Combine script into single file
This is going to be needed to upload the script in the official
scripts repository https://github.com/weechat/scripts

It was potentially doable to craft a build creating it (concatenating
the files in the right order, moving the necessary imports to the top,
renaming some calls to remove the namespaces,...). But the script should
be manageable as a single file so it's simpler to do without that build.
The logic behind the split between files was not clear cut anyway.