Remove short vs full post id distinction
From the cursor mode and the commands, we now directly use the full post
id part.

There will now be a request failure in case we use an invalid post id
but that's fine. Technically this allows interacting with posts not in
the buffer. We can improve the error message reporting later which would
then provide the same as now with an "invalid post id" text.