Fix reaction on posts with files
The last line was since commit b02ee8e072ef8484a9368b9ff2fa9f39fd2c3481
a line with a file displayed. A reaction on the post was replacing it
with last line of the message. Before this change and also commit
c55f4983f6d7fe9447f6de35cf1c0a4fa4bbc251, the reaction on a file only
post was displayed on the empty line above.