Include files in Post build_message
This fixes the issue of files only posts not having a nick displayed.
It also simplify the building of the message in regards to the reactions
line (no need to check whether there are files, and if so need to make a
special case for the last one)
We now only have to add the file_id tags afterwards.